Add tests for redirect URI bug
This commit adds an integration test for the redirect URI fix
from commit fd4d70bcb545bf181c37b52abc7b9da0cad91940.

Since the URI resolving is nested in the controller a few
helpers and mocks were needed to set up the test. These
can be found in the helpers namespace in the test project.

Ref ActiveLogin/ActiveLogin.Authentication#123