Add cancel button to abort BankID authn requests
This commit adds a cancel button to the status view. When clicked
the button triggers a post request to a new endpoint in the
BankIdApiController. The new endpoint then asks the BankID
client to cancel the outstanding authentication request.

If the cancellation is successful control is returned to
the GUI where the status information is hidden and the
form field and buttons for initiating a BankID authn
request are once again visisble.

Ref ActiveLogin/ActiveLogin.Authentication#119