Use QR codes for other device authentication
This commit adds support for QR codes. By default it is only used
for 'other device' authentication in line with the BankID
relying party guidelines.

It uses the auto login functionality since using a QR code follows
the same initialization steps as the auto launch. No PIN is
required and no PIN change is allowed.

A new dependency is added to create the QR codes. It is
a nuget package named 'QRCoder'[1].

Ref ActiveLogin/ActiveLogin.Authentication#19

[1] https://github.com/codebude/QRCoder/