Support certificate identifier when loading from azure
This commit adds support for loading a certificate identifier. It does
so by checking if the identifier is a 'secret' or 'certificate' in
the same manner that the 'SecretIdentifier' and 'CertificateIdentifier'
APIs in core do. By checking for a substring.

This commit does not remove support for fetching the certificates from
the 'secrets' in keyvault to avoid a breaking change.

Ref ActiveLogin/ActiveLogin.Authentication#127