Use cancel return url for samedevice auth
This commit adds a new property 'cancelReturnUrl' that is
used by the samedevice scheme to redirect the user back
to the initial page if they cancel an ongoing auth
process.

Currently it limits the cancelUrl to the samedevice
only but it is possible to allow the cancel url to be
set for the otherdevice scheme as well without too
much hassle.

Ref ActiveLogin/ActiveLogin.Authentication#143