Use QR codes for other device authentication (#138)
* Use QR codes for other device authentication

This commit adds support for QR codes. By default it is only used
for 'other device' authentication in line with the BankID
relying party guidelines.

It uses the auto login functionality since using a QR code follows
the same initialization steps as the auto launch. No PIN is
required and no PIN change is allowed.

A new dependency is added to create the QR codes. It is
a nuget package named 'QRCoder'[1].

Ref ActiveLogin/ActiveLogin.Authentication#19

[1] https://github.com/codebude/QRCoder/

* Fix issues from code review on qr codes

* Scope qr code to statusElement

* No initial src

* Fix style reference

* Use var

* Fix test

* Fix broken test after merge

* Remove unused status wrapper

* Fix review suggestions

* Go back to old layout
* Fix quote usage to follow proper quote style
* Group qr code option and add documentation
* Use bang bangs to check for qr code data
* Remove reduntand styling
* Add new bootstrap styling

* Re-add status wrapper

* Add margin between status and qr code

* Add margin to test case

* Move from 'light' to 'white' theming to match qr code

* Display RFA17B message for qr code fail event

* Add documentation on how to turn off qr codes

* Add semicolon