Use separate package for generating qr codes
This commit breaks out the QRCoder dependency and adds it to a separate
package. It updates the samples and documentation to show how to use the
default qr code.

There is no dependency from 'ActiveLogin.Authentication.BankId.AspNetCore'
to the new 'ActiveLogin.Authentication.BankId.AspNetCore.Qr' package which
means the user will have to import it themselves. They also need to configure
the default qr code implementation if they want to use it.

Ref ActiveLogin/ActiveLogin.Authentication#19