Use separate package for generating qr codes (#155)
* Use separate package for generating qr codes

This commit breaks out the QRCoder dependency and adds it to a separate
package. It updates the samples and documentation to show how to use the
default qr code.

There is no dependency from 'ActiveLogin.Authentication.BankId.AspNetCore'
to the new 'ActiveLogin.Authentication.BankId.AspNetCore.Qr' package which
means the user will have to import it themselves. They also need to configure
the default qr code implementation if they want to use it.

Ref ActiveLogin/ActiveLogin.Authentication#19

* Fix review suggestions for qr coder lib

* Add internal qr generator for qr place holder

* Fix review suggestion for refactoring the qr lib

* Fix typo in swedish QR placeholder image

* Rename placeholder qr generator

* Reorder project file for qr coder

* Revert class name change and set name to correct class for missing qr code config

* Rename from QRCoder to QrCoder...