Provide validated cancellation URL for all BankID cancellations
This commit will swallow any request error between the API and the
BankID service when the user tries to cancel. When a user cancels,
they only have one way back to restart the authentication process
and that is through the cancellation button.

We still keep the validation for order reference and cancel url
though since we want to protect against bad requests coming in
and potential redirection attacks.

Ref ActiveLogin/ActiveLogin.Authentication#184