Add extension point for refresh behavior on iOS
This patch adds an extension point for controlling the refresh browser
behaviour when returning from the BankID app.

The problem we are trying to solve is that when starting the flow from
a native iOS app using an embedded view like
ASWebAuthenticationSession, we do not want to refresh the browser.
With the new IReloadPageOnReturnFromBankIdApp interface, we can simply
disable the refresh.